+++ lib/src/commands/commands.dart
@@ -1,3 +1,4 @@
+export 'cache.dart';
 export 'get.dart';
 export 'init.dart';
 export 'make.dart';